+++ packages/clients/diplan/src/polar-client.ts
@@ -45,6 +45,7 @@ polarCore.addPlugins([
       },
       {
         plugin: Draw({
+          toastAction: 'plugin/toast/addToast',
           addLoading: 'plugin/loadingIndicator/addLoadingKey',
           removeLoading: 'plugin/loadingIndicator/removeLoadingKey',
         }),