+++ packages/clients/diplan/example/config.js
@@ -111,6 +111,14 @@ export default {
   },
   draw: {
     enableOptions: true,
+    lassos: [
+      {
+        id: flurstuecke,
+      },
+      {
+        id: xplanwfs,
+      },
+    ],
     measureOptions: {
       metres: true,
       kilometres: true,